+++ src/main/java/ru/miroque/personal/profile/model/concept/service/ServiceKnowledge.java
@@ -8,7 +8,7 @@ import java.util.Collection;
 
 public interface ServiceKnowledge {
 
-	Collection<Knowledge> findByName(String name) throws XPathExpressionException;
+	Collection<Knowledge> findByName(String value) throws XPathExpressionException;
 
 	void set(Knowledge item) throws ExceptionNotPersisted;